tr///c: handle len(replacement charlist) > 32767
RT #132608

In the non-utf8 case, the /c (complement) flag to tr adds an implied
\x{100}-\x{7fffffff} range to the search charlist. If the replacement list
contains more chars than are paired with the 0-255 part of the search
list, then the excess chars are stored in an extended part of the table.
The excess char count was being stored as a short, which caused problems
if the replacement list contained more than 32767 excess chars: either
substituting the wrong char, or substituting for a char located up to
0xffff bytes in memory before the real translation table.

So change it to SSize_t.

Note that this is only a problem when the search and replacement charlists
are non-utf8, the replacement list contains around 0x8000+ entries, and
where the string being translated is utf8 with at least one codepoint >=
U+8000.

+++ op.c
@@ -6702,7 +6702,7 @@ S_pmtrans(pTHX_ OP *o, OP *expr, OP *repl)
                         o->op_private |= OPpTRANS_IDENTICAL;
                 }
 	    }
-            extbl->excess_len  = (short)excess;      /* excess char count */
+            extbl->excess_len  = excess;      /* excess char count */
             extbl->repeat_char = (short)repeat_char; /* repeated replace char */
 	}
     }
